move files into single_event subfolder
