+++ example/GW150914_heterodyne.py
@@ -1,8 +1,8 @@
 import time
 from jimgw.jim import Jim
-from jimgw.detector import H1, L1
-from jimgw.likelihood import HeterodynedTransientLikelihoodFD, TransientLikelihoodFD
-from jimgw.waveform import RippleIMRPhenomD
+from jimgw.single_event.detector import H1, L1
+from jimgw.single_event.likelihood import HeterodynedTransientLikelihoodFD, TransientLikelihoodFD
+from jimgw.single_event.waveform import RippleIMRPhenomD
 from jimgw.prior import Uniform, Composite
 import jax.numpy as jnp
 import jax
