+++ src/jimgw/single_event/wave.py
