+++ src/jimgw/single_event/waveform.py
